Clean the code with metrics exporter
clean the code

Add limit resource requirements for lifecycle container

Increase hard limit for alpine container
